Refactor SessionsController
- Remove `OauthLogin` and `SessionPersonCreator` and fold them into
  `SessionsController` - they are not reused anywhere and just add
  unnecessary layers of indirection
- Remove `#create_person` - this is not used anywhere
