+++ spec/controllers/sessions_controller_spec.rb
@@ -1,8 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe SessionsController, type: :controller do
-  it_behaves_like 'session_person_creatable'
-
   let!(:person) { create(:person, given_name: 'John', surname: 'Doe', email: 'john.doe@digital.justice.gov.uk', ditsso_user_id: 'deadbeef') }
 
   let(:john_doe_omniauth_hash) do
@@ -100,28 +98,6 @@ RSpec.describe SessionsController, type: :controller do
           expect(response).to redirect_to edit_person_path(Person.find_by(ditsso_user_id: 'dec0de'), page_title: "Create profile")
         end
       end
-    end
-  end
-
-  describe 'POST create_person' do
-    let(:person_params) do
-      {
-        person: {
-          email: 'fred.bloggs@digital.justice.gov.uk',
-          given_name: 'Fred',
-          surname: 'Bloggs'
-        }
-      }
-    end
-
-    it 'creates the new person' do
-      expect { post :create_person, params: person_params }.to change Person, :count
-    end
-
-    it 'redirects to the person\'s profile edit page, ignoring desired path' do
-      request.session[:desired_path] = '/search'
-      post :create_person, params: person_params
-      expect(response).to redirect_to edit_person_path(Person.find_by(email: 'fred.bloggs@digital.justice.gov.uk'), page_title: "Create profile")
     end
   end
 end
